refactor bloc, implement daily weather screen and bind it with data

+++ lib/features/view_weather/data/dto/daily_weather_dto.dart
@@ -39,7 +39,7 @@ class DailyWeatherDto {
 extension DailyWeatherMapper on DailyWeatherDto {
   DailyWeather toDailyWeather() => DailyWeather(
         weather: weather.toWeather(),
-        windSpeed: windSpeed,
+        windSpeed: windSpeed.toInt(),
         maxTemp: maxTemp.toInt(),
         minTemp: minTemp.toInt(),
         humidity: humidity,
